# PERF: Skip _shallow_copy for RangeIndex._join_empty where other is RangeIndex