- [ ] closes #57852 (Replace xxxx with the GitHub issue number)


```python
In [1]: from pandas import *; import numpy as np
+ /opt/miniconda3/envs/pandas-dev/bin/ninja
[1/1] Generating write_version_file with a custom command

In [2]: df = DataFrame({"A": np.arange(100_000)})

In [3]: df_empty = DataFrame(columns=["B", "C"], dtype="int64")

In [4]: %timeit df_empty.join(df, how="inner")
369 µs ± 4.17 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)  # main

In [4]: %timeit df_empty.join(df, how="inner")
126 µs ± 1.74 µs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)  # PR
```